타임딜 주문 기능 추가

+++ src/main/java/com/specialpriceshop/item/dto/StockResponse.java
@@ -1,8 +1,9 @@
 package com.specialpriceshop.item.dto;
 
 import com.specialpriceshop.item.domain.Stock;
+import java.math.BigDecimal;
 
-public record StockResponse(Long stockId, String optionName, double addPrice) {
+public record StockResponse(Long stockId, String optionName, BigDecimal addPrice) {
 
     public static StockResponse of(Stock entity) {
         return new StockResponse(entity.getId(), entity.getOptionName(), entity.getAddPrice());
